+++ apps/server/src/modules/common/warehouses/warehouses.module.ts
@@ -1,9 +1,9 @@
 import { Module } from "@nestjs/common";
+import { commonMasterProviders } from "../infrastructure/common-master.providers";
 import { WarehousesController } from "./warehouses.controller";
-import { WarehousesRepository } from "./warehouses.repository";
 
 @Module({
   controllers: [WarehousesController],
-  providers: [WarehousesRepository],
+  providers: [...commonMasterProviders],
 })
 export class WarehousesModule {}